Send uploaded file paths to the grading script.
